[#140] Merge reset password and invitation
